requirements.txt
python-telegram-bot
azure-cognitiveservices-vision-computervision
opencv-python
numpy
Pillow
python-dotenv